refactor: intro test folder
