+++ packages/utils/tsconfig.lib.json
@@ -10,6 +10,7 @@
     "vite.config.ts",
     "src/**/*.spec.ts",
     "src/**/*.test.ts",
-    "src/**/*.mock.ts"
+    "src/**/*.mock.ts",
+    "test/**/*.ts"
   ]
 }
