+++ packages/utils/tsconfig.spec.json
@@ -6,6 +6,7 @@
   },
   "include": [
     "vite.config.ts",
+    "test/**/*.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
     "src/**/*.test.tsx",
